requirement.txt
certifi==2018.8.24
chardet==3.0.4
Django==1.11.15
django-extensions==2.1.0
django-js-asset==1.1.0
django-kronos==1.0
get==1.0.3
idna==2.6
post==1.0.2
public==1.0.3
python-crontab==2.3.5
python-dateutil==2.7.3
pytz==2018.5
query-string==1.0.2
request==1.0.2
requests==2.18.4
six==1.11.0
urllib3==1.22